Possibilité de supprimé une annonce sans photos

+++ AstroLogement/src/Controller/LocationController.php
@@ -128,12 +128,14 @@ final class LocationController extends AbstractController{
     public function delete(Request $request, Location $location, EntityManagerInterface $entityManager): Response
     {
         if ($this->isCsrfTokenValid('delete'.$location->getId(), $request->getPayload()->getString('_token'))) {
-            foreach ($location->getPhotos() as $photo) {
-                $photoPath = $this->getParameter('kernel.project_dir') . "/public/" . $photo;
-                if (file_exists($photoPath)) {
-                    unlink($photoPath);
+            if($location->getPhotos()){    
+                foreach ($location->getPhotos() as $photo) {
+                    $photoPath = $this->getParameter('kernel.project_dir') . "/public/" . $photo;
+                    if (file_exists($photoPath)) {
+                        unlink($photoPath);
+                    }
                 }
-            }
+            } 
             $entityManager->remove($location);
             $entityManager->flush();
         }
